MVC 6 - Creating a Layout
Added a layout view under the Views folder in a folder called Shared.
The layout view is commonly named “_layout.cshtml”. In order for MVC to
know about what its called, you define it in a “_ViewStart.cshtml” file
under the Views folder.  You then use @RenderBody() in the layout to
define where the content of each different view is injected.